修复中。。。

+++ chatly2/src/pages/chat/friend/detail.jsx
@@ -3,9 +3,11 @@ import { useNavigate, useLocation } from "react-router";
 import { useWinSize } from 'hooks';
 import { useHttpClient2, useApiBase } from 'hooks/http';
 import { db } from 'hooks/db';
-import { useRequest } from 'ahooks';
 import { Avatar, Button, Divider, Heading, YBox, XBox } from 'components/flutter';
 import { InfoTile } from 'components/chat';
+import { useMutation } from '@tanstack/react-query';
+
+
 
 export function Detail() {
   const navigate = useNavigate();
@@ -27,30 +29,37 @@ export function Detail() {
     return `${String(apiBase || "").replace(/\/+$/, "")}/imgs/${String(friend.avatar_url).replace(/^\/+/, "")}`;
   }, [apiBase, friend?.avatar_url]);
 
-  const { runAsync: delFid } = useRequest(
-    (id) => {
-      return http2.requestBodyJson('DELETE', { id }).then((results) => {
-        if (!results) return;
-        return db.table('friends').get(id).then((row) => {
-          return Promise.all([
-            db.table('message').where('uid').equals(row?.uid).delete(),
-            db.table('friends').delete(id),
-          ]);
-        });
-      });
-    },
-    { manual: true }
-  );
 
-  const { runAsync: updRemark } = useRequest(
-    (id, remark) => {
-      return http2.requestBodyJson('PATCH', { id, remark }).then((results) => {
-        if (!results) return;
-        return db.table('friends').update(id, { remark });
-      });
+  const { mutateAsync: delFid } = useMutation({
+    mutationFn: async (id) => {
+      if (!id) {
+        return;
+      }
+      const results = await http2.requestBodyJson('DELETE', { id });
+      if (!results) return;
+      const row = await db.table('friends').get(id);
+
+      await Promise.all([
+        db.table('message').where('uid').equals(row?.uid).delete(),
+        db.table('friends').delete(id),
+      ]);
+
+      return 'ok';
     },
-    { manual: true }
-  );
+  });
+
+
+
+  const { mutateAsync: updRemark } = useMutation({
+    mutationFn: async ({ id, remark }) => {
+      if (!id) return;
+      const results = await http2.requestBodyJson('PATCH', { id, remark });
+      if (!results) return;
+      await db.table('friends').update(id, { remark });
+      return 'ok';
+    },
+  });
+
 
   function openMsgWindow(friend) {
     if (!friend?.id) return;
@@ -107,7 +116,7 @@ export function Detail() {
             value={friend?.remark}
             onConfirm={(remark) => {
               setFriend((p) => ({ ...p, remark }));
-              updRemark(friend?.id, remark);
+              updRemark({ id: friend?.id, remark });
             }}
           />
         </YBox.Segment>
@@ -118,8 +127,8 @@ export function Detail() {
               label="发起聊天"
               onPressed={() => openMsgWindow(friend)}
               style={{
-                background: 'var(--accent-color)',
                 color: '#fff',
+                background: '#4d4dff',
                 border: 'none',
               }}
             />
